pyproject.toml
[project]
name = "medical-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "huggingface-hub>=1.1.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.10",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
]
